requirements.txt
fastnumbers>=3.2
fastybird-metadata>=0.24
kink>=0.6
setuptools>=57.4
sqlalchemy[mypy]>=1.4
pymysql>=1.0